+++ resonance/nonlinear_systems.py
@@ -4,7 +4,7 @@ import numpy as np
 import scipy as sp
 import scipy.integrate  # scipy doesn't import automatically
 
-from .systems import _System
+from .system import _System
 
 
 class NonLinearSystem(_System):
